ORD-758 refactor to shared page objects, update dependencies
