+++ project/Dependencies.scala
@@ -3,7 +3,7 @@ import sbt.*
 object Dependencies {
 
   val test: Seq[ModuleID] = Seq(
-    "uk.gov.hmrc"                %% "ui-test-runner"          % "0.44.0",
+    "uk.gov.hmrc"                %% "ui-test-runner"          % "0.54.0",
     "io.cucumber"                %% "cucumber-scala"          % "8.25.1",
     "io.cucumber"                %  "cucumber-junit"          % "7.19.0",
     "junit"                      %  "junit"                   % "4.13.2",
