+++ project/plugins.sbt
@@ -2,4 +2,4 @@ resolvers += MavenRepository("HMRC-open-artefacts-maven2", "https://open.artefac
 resolvers += Resolver.url("HMRC-open-artefacts-ivy2", url("https://open.artefacts.tax.service.gov.uk/ivy2"))(Resolver.ivyStylePatterns)
 
 addSbtPlugin("uk.gov.hmrc" % "sbt-auto-build" %  "3.24.0")
-addSbtPlugin("uk.gov.hmrc" % "sbt-test-report" % "1.10.0")
+addSbtPlugin("uk.gov.hmrc" % "sbt-test-report" % "1.12.0")
