+++ run_tests.sh
@@ -2,4 +2,4 @@
 ENV=${1:-local}
 BROWSER=${2:-chrome}
 
-sbt clean -Dbrowser="${BROWSER}" -Denvironment="${ENV}" -Dbrowser.option.headless=false "testOnly uk.gov.hmrc.test.ui.cucumber.runner.Runner" testReport
+sbt clean -Dbrowser="${BROWSER}" -Denvironment="${ENV}" -Dbrowser.option.headless=true -Dbrowser.usePreviousVersion=true "testOnly uk.gov.hmrc.test.ui.specs* -- -n uk.gov.hmrc.test.ui.specs.tags.MVP" testReport
