+++ src/test/scala/uk/gov/hmrc/test/ui/pages/AuthLoginPage.scala
@@ -18,49 +18,49 @@ package uk.gov.hmrc.test.ui.pages
 
 import org.openqa.selenium.support.ui.Select
 import org.openqa.selenium.{By, WebElement}
+import uk.gov.hmrc.selenium.webdriver.Driver
 import uk.gov.hmrc.test.ui.conf.TestConfiguration
-import uk.gov.hmrc.test.ui.driver.BrowserDriver
 
-object AuthLoginPage extends BrowserDriver {
+object AuthLoginPage {
   val url: String = TestConfiguration.url("auth-login-stub") + "/gg-sign-in"
 
-  def getRedirectUrl: WebElement = driver.findElement(By.name("redirectionUrl"))
+  def getRedirectUrl: WebElement = Driver.instance.findElement(By.name("redirectionUrl"))
 
   def selectCredentialStrength(credentialStrength: String): Unit = {
-    val select = new Select(driver.findElement(By.name("credentialStrength")))
+    val select = new Select(Driver.instance.findElement(By.name("credentialStrength")))
     select.selectByVisibleText(credentialStrength)
   }
 
   def selectConfidenceLevel(confidenceLevel: String): Unit = {
-    val select = new Select(driver.findElement(By.name("confidenceLevel")))
+    val select = new Select(Driver.instance.findElement(By.name("confidenceLevel")))
     select.selectByVisibleText(confidenceLevel)
   }
 
   def selectAffinityGroup(affinityGroup: String): Unit = {
-    val select = new Select(driver.findElement(By.name("affinityGroup")))
+    val select = new Select(Driver.instance.findElement(By.name("affinityGroup")))
     select.selectByVisibleText(affinityGroup)
   }
 
-  def getNino: WebElement = driver.findElement(By.name("nino"))
+  def getNino: WebElement = Driver.instance.findElement(By.name("nino"))
 
-  def getEnrolmentKeyOne: WebElement = driver.findElement(By.name("enrolment[0].name"))
+  def getEnrolmentKeyOne: WebElement = Driver.instance.findElement(By.name("enrolment[0].name"))
 
-  def getIdentifierNameOne: WebElement = driver.findElement(By.name("enrolment[0].taxIdentifier[0].name"))
+  def getIdentifierNameOne: WebElement = Driver.instance.findElement(By.name("enrolment[0].taxIdentifier[0].name"))
 
-  def getIdentifierValueOne: WebElement = driver.findElement(By.name("enrolment[0].taxIdentifier[0].value"))
+  def getIdentifierValueOne: WebElement = Driver.instance.findElement(By.name("enrolment[0].taxIdentifier[0].value"))
 
-  def getEnrolmentKeyTwo: WebElement = driver.findElement(By.name("enrolment[1].name"))
+  def getEnrolmentKeyTwo: WebElement = Driver.instance.findElement(By.name("enrolment[1].name"))
 
-  def getIdentifierNameTwo: WebElement = driver.findElement(By.name("enrolment[1].taxIdentifier[0].name"))
+  def getIdentifierNameTwo: WebElement = Driver.instance.findElement(By.name("enrolment[1].taxIdentifier[0].name"))
 
-  def getIdentifierValueTwo: WebElement = driver.findElement(By.name("enrolment[1].taxIdentifier[0].value"))
+  def getIdentifierValueTwo: WebElement = Driver.instance.findElement(By.name("enrolment[1].taxIdentifier[0].value"))
 
-  def getDelegatedEnrolmentKey: WebElement = driver.findElement(By.name("delegatedEnrolment[0].key"))
+  def getDelegatedEnrolmentKey: WebElement = Driver.instance.findElement(By.name("delegatedEnrolment[0].key"))
 
-  def getDelegatedIdentifierName: WebElement = driver.findElement(By.name("delegatedEnrolment[0].taxIdentifier[0].name"))
+  def getDelegatedIdentifierName: WebElement = Driver.instance.findElement(By.name("delegatedEnrolment[0].taxIdentifier[0].name"))
 
-  def getDelegatedIdentifierValue: WebElement = driver.findElement(By.name("delegatedEnrolment[0].taxIdentifier[0].value"))
+  def getDelegatedIdentifierValue: WebElement = Driver.instance.findElement(By.name("delegatedEnrolment[0].taxIdentifier[0].value"))
 
-  def getDelegatedAuthRule: WebElement = driver.findElement(By.name("delegatedEnrolment[0].delegatedAuthRule"))
+  def getDelegatedAuthRule: WebElement = Driver.instance.findElement(By.name("delegatedEnrolment[0].delegatedAuthRule"))
 
 }
