--- src/test/scala/uk/gov/hmrc/test/ui/scalatestpages/CommonElementsNew.scala
@@ -1,932 +1,0 @@
-/*
- * Copyright 2023 HM Revenue & Customs
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package uk.gov.hmrc.test.ui.pages
-
-import org.openqa.selenium.By
-import org.scalatest.matchers.should.Matchers
-
-trait CommonElementsNew extends BasePageNew with Matchers {
-
-  def load(key: String): By =
-    elements(key)
-
-  private[pages] val commonMapping: List[(String, By)] = List(
-    //Common Elements
-    "yes"                       -> By.id("value"),
-    "no"                        -> By.id("value-no"),
-    "Yes"                       -> By.xpath("//*[contains(text(),'Yes')]"),
-    "No"                        -> By.xpath("//*[contains(text(),'No')]"),
-    "Continue"                  -> By.xpath("//*[contains(text(),'Continue')]"),
-    "save and continue"         -> By.cssSelector("#main-content > div > div > form > button"),
-    "continue"                  -> By.id("continue"),
-    "confirm"                   -> By.xpath("//*[contains(text(),'Confirm')]"),
-    "back"                      -> By.cssSelector("a.govuk-back-link"),
-    "remove account"            -> By.cssSelector("#main-content > div > div > div.govuk-form-group > dl > div:nth-child(1) > dd > ul > li:nth-child(2) > a"),
-    "change"                    -> By.cssSelector("#main-content > div > div > div.govuk-form-group > ul > li > span.hmrc-add-to-a-list__change > a"),
-    "return to overview"        -> By.id("returnToOverviewPageBtn"),
-    "remove"                    -> By.id("remove"),
-    "Yes for gateway question"  -> By.xpath("//dd[contains(text(),'Yes')]"),
-    "No for gateway question"   -> By.xpath("//dd[contains(text(),'No')]"),
-    "amount"                    -> By.id("amount"),
-    "cancel"                    -> By.xpath("//*[contains(text(),'Cancel')]"),
-    "agentCheckboxV&C"          -> By.xpath("//*[@id='Agent']"),
-    "vcLoginButton"             -> By.xpath("//*[@class='govuk-button']"),
-
-    //Tailoring Elements
-    "addSections"   -> By.id("addSectionsLink"),
-    "addDividend"   -> By.id("dividends"),
-    "addInterest"   -> By.id("interest"),
-    "addGiftAid"    -> By.id("gift-aid"),
-    "addEmployment" -> By.id("employment"),
-    "addCisSection" -> By.id("cis"),
-    "addPension"    -> By.id("pensions"),
-    "addGains"      -> By.id("gains"),
-    "addProperty"   -> By.id("property"),
-    "addSelfEmployment" -> By.id("self-employment"),
-
-
-    //Common Task List Elements
-    "UK residence status"                                 -> By.cssSelector("#main-content > div > div > ul:nth-child(5) > li:nth-child(1) > div.govuk-task-list__name-and-hint > a"),
-    "Foster carer status"                                 -> By.cssSelector("#main-content > div > div > ul:nth-child(5) > li:nth-child(2) > div.govuk-task-list__name-and-hint > a"),
-    "Donations using Gift Aid task list"                  -> By.className("DonationsUsingGiftAid"),
-    "Gifts of land or property task list"                 -> By.className("GiftsOfLandOrProperty"),
-    "Gifts of shares or securities task list"             -> By.className("GiftsOfShares"),
-    "Gifts of overseas charities task list"               -> By.className("GiftsToOverseas"),
-    "PAYE employment (including student loans) task list" -> By.className("PayeEmployment"),
-    "Construction Industry Schemes deductions"            -> By.className("CIS"),
-    "Review Employment and support allowance claims"      -> By.className("ESA"),
-    "Review Jobseeker’s Allowance claims"                 -> By.className("JSA"),
-    "State Pension1"                                      -> By.className("StatePension"),
-    "Other UK Pensions1"                                  -> By.className("OtherUkPensions"),
-    "Unauthorised payments1"                              -> By.className("UnauthorisedPayments"),
-    "Short service refunds1"                              -> By.className("ShortServiceRefunds"),
-    "Income from overseas pensions1"                      -> By.className("IncomeFromOverseas"),
-    "Payments into UK pensions1"                          -> By.className("PaymentsIntoUk"),
-    "Annual allowances1"                                  -> By.className("AnnualAllowances"),
-    "Payments into overseas pensions"                     -> By.className("PaymentsIntoOverseas"),
-    "Overseas transfer charge1"                           -> By.className("OverseasTransfer"),
-    "Banks and building societies task list"                       -> By.className("BanksAndBuilding"),
-    "Trust fund bond or bonds task list"                            -> By.className("TrustFundBond"),
-    "Gilt-edged or accrued income securities task list"            -> By.className("GiltEdged"),
-    "Cash dividends from UK stocks and shares task list"           -> By.className("CashDividends"),
-    "Stock dividends from UK companies task list"                  -> By.className("StockDividends"),
-    "Dividends from unit trusts or investment companies task list" -> By.className("DividendsFromUnitTrusts"),
-    "Free or redeemable shares task list"                          -> By.className("FreeRedeemableShares"),
-    "Close company loans written off or released task list"        -> By.className("CloseCompanyLoans"),
-    "Life Insurance common task list"                     -> By.className("LifeInsurance"),
-    "Life Annuity common task list"                       -> By.className("LifeAnnuity"),
-    "Capital Redemption common task list"                 -> By.className("CapitalRedemption"),
-    "Voided ISA(a policy cancelled by your ISA manager) common task list" -> By.className("VoidedISA"),
-    "UK Foreign Property" -> By.className("UkForeignProperty"),
-    "UKProperty" -> By.className("UkProperty"),
-    "ForeignProperty" -> By.className("ForeignProperty"),
-
-    //Dividends Elements
-    "dividends"                                                  -> By.id("dividends_link"),
-    "UK dividends amount"                                        -> By.id("amount"),
-    "UK other dividends amount"                                  -> By.id("amount"),
-    "change dividends from the uk summary"                       -> By.cssSelector("#main-content > div > div > dl > div:nth-child(1) > dd.govuk-summary-list__actions > a"),
-    "change uk dividends amount"                                 -> By.cssSelector("#main-content > div > div > dl > div:nth-child(2) > dd.govuk-summary-list__actions > a"),
-    "change other dividends amount"                              -> By.cssSelector("#main-content > div > div > dl > div:nth-child(3) > dd.govuk-summary-list__actions > a"),
-    "change stock dividends"                                     -> By.cssSelector("#main-content > div > div > dl > div:nth-child(4) > dd.govuk-summary-list__actions > a"),
-    "change value of dividends from UK companies"                -> By.xpath("//*[contains(text(),'amount of dividends from UK companies')]/parent::a"),
-    "change value of stock dividends"                            -> By.xpath("//*[contains(text(),'Value of stock dividends')]/parent::div/dd[2]/a"),
-    "change value of free or redeemable shares"                  -> By.xpath("//*[contains(text(),' value of free or redeemable shares')]/parent::a"),
-    "change value of unit trusts or investment companies"        -> By.xpath("//*[contains(text(),' amount of dividends from unit trusts or investment companies')]/parent::a"),
-    "change value of close company loan written off or released" -> By.xpath("//*[contains(text(),'Close company loan written off or released')]/parent::div/dd[2]/a"),
-
-    //Interest Elements
-    "Interest from savings and securities"                  -> By.id("interest_link"),
-    "UK Interest"                                           -> By.xpath("//*[contains(text(),'UK Interest')]"),
-    "Interest from gilt-edged or accrued income securities" -> By.cssSelector("#main-content > div > div > ol > li:nth-child(2) > span.app-task-list__task-name > a"),
-    "UK untaxed interest amount earned"           -> By.id("untaxedAmount"),
-    "UK taxed interest amount earned"             -> By.id("taxedAmount"),
-    "UK untaxed interest updated amount earned"   -> By.id("amount"),
-    "UK taxed interest updated amount earned"     -> By.id("amount"),
-    "UK untaxed interest account name"            -> By.id("untaxedAccountName"),
-    "UK taxed interest account name"              -> By.id("taxedAccountName"),
-    "first account change"                        -> By.cssSelector("#main-content > div > div > dl > div:nth-child(1) > dd > ul > li:nth-child(1) > a"),
-    "second account change"                       -> By.cssSelector("#main-content > div > div > dl > div:nth-child(2) > dd > ul > li:nth-child(1) > a"),
-    "change interest from the uk summary"         -> By.cssSelector("#main-content > div > div > dl > div:nth-child(1) > dd.govuk-summary-list__actions > a"),
-    "change uk untaxed interest accounts summary" -> By.cssSelector("#main-content > div > div > dl > div:nth-child(2) > dd.govuk-summary-list__actions > a"),
-    "change uk taxed interest accounts summary"   -> By.cssSelector("#main-content > div > div > dl > div:nth-child(3) > dd.govuk-summary-list__actions > a"),
-    "change amount of interest"                   -> By.xpath("//*[contains(text(),' Change amount of interest')]/parent::a"),
-    "remove interest"                             -> By.cssSelector("#main-content > div > div > dl > div:nth-child(1) > dd > ul > li:nth-child(2) > a"),
-    "first"                                       -> By.id("value"),
-    "second"                                      -> By.id("value-2"),
-    "third"                                       -> By.id("value-3"),
-    "fourth"                                      -> By.id("value-4"),
-    "fifth"                                       -> By.id("value-5"),
-
-    //Employment Elements
-    "employment"                   -> By.id("employment_link"),
-    "employment details"           -> By.linkText("Employment details"),
-    "view employer"                -> By.cssSelector("#main-content > div > div > dl:nth-of-type(1) > div > dd > a"),
-    "benefits"                     -> By.linkText("Benefits"),
-    "employment benefits"          -> By.id("employment-benefits_link"),
-    "expenses"                     -> By.linkText("Expenses"),
-    "view expenses"                -> By.cssSelector("#main-content > div > div > dl:nth-of-type(2) > div > dd > a"),
-    "add expenses"                 -> By.id("add-expenses"),
-    "change expenses"              -> By.cssSelector("#main-content > div > div > dl:nth-child(10) > div > dd > ul > li:nth-child(1) > a"),
-    "view first employer"          -> By.cssSelector("#main-content > div > div > dl:nth-of-type(1) > div:nth-child(1) > dd > a"),
-    "change employer"              -> By.cssSelector("#main-content > div > div > dl:nth-child(6) > div > dd > ul > li:nth-child(1) > a"),
-    "return to employment summary" -> By.id("returnToEmploymentSummaryBtn"),
-    "return to employer"           -> By.id("returnToEmployerBtn"),
-    "Clickable employment"         -> By.id("employment_link"),
-    "Employer name"                -> By.id("name"),
-    "PAYE reference"               -> By.id("payeRef"),
-    "Payroll id"                   -> By.id("payrollId"),
-    "Day"                          -> By.id("amount-day"),
-    "Month"                        -> By.id("amount-month"),
-    "Year"                         -> By.id("amount-year"),
-    "Start Day"                    -> By.id("startAmount-day"),
-    "Start Month"                  -> By.id("startAmount-month"),
-    "Start Year"                   -> By.id("startAmount-year"),
-    "End Day"                      -> By.id("endAmount-day"),
-    "End Month"                    -> By.id("endAmount-month"),
-    "End Year"                     -> By.id("endAmount-year"),
-    "Pay"                          -> By.id("amount"),
-    "Tax"                          -> By.id("amount"),
-    "amount"                       -> By.id("amount"),
-    "change employment" -> By.cssSelector("#main-content > div > div > dl > div > dd.govuk-summary-list__actions > ul > li:nth-child(1) > a"),
-    "first change employment" -> By.cssSelector("#main-content > div > div > dl:nth-child(6) > div:nth-child(1) > dd > ul > li:nth-child(1) > a"),
-    "second change employment" -> By.cssSelector("#main-content > div > div > dl:nth-child(6) > div:nth-child(2) > dd > ul > li:nth-child(1) > a"),
-    "third change employment" -> By.cssSelector("#main-content > div > div > dl:nth-child(6) > div:nth-child(3) > dd > ul > li:nth-child(1) > a"),
-    "change Employer" -> By.cssSelector("#main-content > div > div > dl > div:nth-child(1) > dd.govuk-summary-list__actions > a > span:nth-child(1)"),
-    "change start date" -> By.cssSelector("#main-content > div > div > dl > div:nth-child(2) > dd.govuk-summary-list__actions > a > span:nth-child(1)"),
-    "change still working reference" -> By.cssSelector("#main-content > div > div > dl > div:nth-child(3) > dd.govuk-summary-list__actions > a > span:nth-child(1)"),
-    "change end date" -> By.cssSelector("#main-content > div > div > dl > div:nth-child(4) > dd.govuk-summary-list__actions > a > span:nth-child(1)"),
-    "change PAYE reference" -> By.cssSelector("#main-content > div > div > dl > div:nth-child(5) > dd.govuk-summary-list__actions > a > span:nth-child(1)"),
-    "change payroll id" -> By.cssSelector("#main-content > div > div > dl > div:nth-child(6) > dd.govuk-summary-list__actions > a > span:nth-child(1)"),
-    "change pay received" -> By.cssSelector("#main-content > div > div > dl > div:nth-child(7) > dd.govuk-summary-list__actions > a > span:nth-child(1)"),
-    "change tax taken from pay" -> By.cssSelector("#main-content > div > div > dl > div:nth-child(8) > dd.govuk-summary-list__actions > a > span:nth-child(1)"),
-    "change off payroll" -> By.cssSelector("#main-content > div > div > dl > div:nth-child(9) > dd.govuk-summary-list__actions > a > span:nth-child(1)"),
-    "Employments list item" -> By.cssSelector("#employer-h2"),
-    "Expenses list item" -> By.cssSelector("#expense-h2"),
-    "remove employment" -> By.cssSelector("#main-content > div > div > dl:nth-child(6) > div > dd > ul > li:nth-child(2) > a"),
-    "third remove employment" -> By.cssSelector("#main-content > div > div > dl:nth-child(6) > div:nth-child(3) > dd > ul > li:nth-child(2) > a"),
-    "remove expenses" -> By.cssSelector("#main-content > div > div > dl:nth-child(8) > div > dd > ul > li:nth-child(2) > a"),
-    "Remove expenses" -> By.cssSelector("#remove-expenses-button-id"),
-    "change expenses section" -> By.cssSelector("#main-content > div > div > div:nth-child(6) > ul > li > span.hmrc-add-to-a-list__change > a"),
-    "Add an employer" -> By.xpath("//a[contains(text(),'Add an employer')]"),
-    "Add another employer" -> By.linkText("Add another employer"),
-    "Remove employer" -> By.id("remove-employer-button-id"),
-    "student loans" -> By.id("student-loans_link"),
-    "student loan repayments" -> By.cssSelector("#studentLoans"),
-    "postgraduate loan repayments" -> By.cssSelector("#studentLoans-2"),
-    "no repayments" -> By.cssSelector("#studentLoans-4"),
-    "change student loan repayments" -> By.cssSelector("#main-content > div > div > dl > div:nth-child(1) > dd.govuk-summary-list__actions > a > span:nth-child(1)"),
-    "student loans amount change" -> By.cssSelector("#main-content > div > div > dl > div:nth-child(2) > dd.govuk-summary-list__actions > a > span:nth-child(1)"),
-    "postgraduate loans amount change" -> By.cssSelector("#main-content > div > div > dl > div:nth-child(3) > dd.govuk-summary-list__actions > a > span:nth-child(1)"),
-    "cannot update" -> By.cssSelector("#main-content > div > div > div.govuk-inset-text"),
-    "PAYE employment change" -> By.cssSelector("#main-content > div > div > dl:nth-child(3) > div > dd.govuk-summary-list__actions.govuk-\\!-width-one-third > a > span:nth-child(1)"),
-    "Employment remove confirm" -> By.id("remove-employer-button-id"),
-    "Taxable lump sums and certain income" -> By.id("taxable-lump-sums-and-certain-income_link"),
-    "Taxable lump sum page" -> By.id("page-heading"),
-    "Taxable lump sum button" -> By.id("AddALumpSum"),
-    "Taxable lump sum amount page" -> By.id("page-heading"),
-    "Taxable lumps su amount" -> By.id("amount"),
-    "Taxable lump sum continue button" -> By.id("continue"),
-    "Save and continue" -> By.id("saveAndContinueBtn"),
-
-    //GiftAid Elements
-    "GiftAid"                                             -> By.id("giftAid_link"),
-    "Amount donated using Gift Aid"                       -> By.id("amount"),
-    "Amount donated as one-off payments"                  -> By.id("amount"),
-    "Amount donated to overseas charities using Gift Aid" -> By.id("amount"),
-    "Name of overseas charity"                            -> By.id("name"),
-    "Amount to add to last tax year"                      -> By.id("amount"),
-    "Amount to add to this tax year"                      -> By.id("amount"),
-    "Total value of qualifying shares or securities"      -> By.id("amount"),
-    "Value of land or property"                           -> By.id("amount"),
-    "Value of qualifying shares, securities, land or property donated to overseas charities"  -> By.id("amount"),
-    "Name of overseas charity you donated shares, securities, land or property to"            -> By.id("name"),
-    "Name of overseas charity your client donated shares, securities, land or property to"    -> By.id("name"),
-    "change donations to charity summary" -> By.cssSelector("#main-content > div > div > dl > div:nth-child(1) > dd.govuk-summary-list__actions.govuk-\\!-width-one-third > a"),
-    "change amount donated to charity using Gift Aid" -> By.cssSelector("#main-content > div > div > dl > div:nth-child(2) > dd.govuk-summary-list__actions.govuk-\\!-width-one-third > a"),
-    "change amount of one-off donations to charity using Gift Aid" -> By.cssSelector("#main-content > div > div > dl > div:nth-child(3) > dd.govuk-summary-list__actions.govuk-\\!-width-one-third > a"),
-    "change amount donated to overseas charities by using Gift Aid" -> By.cssSelector("#main-content > div > div > dl > div:nth-child(4) > dd.govuk-summary-list__actions.govuk-\\!-width-one-third > a"),
-    "change overseas charities names" -> By.cssSelector("#main-content > div > div > dl > div:nth-child(5) > dd.govuk-summary-list__actions.govuk-\\!-width-one-third > a"),
-    "change amount of donation to add to last tax year" -> By.cssSelector("#main-content > div > div > dl > div:nth-child(6) > dd.govuk-summary-list__actions.govuk-\\!-width-one-third > a"),
-    "change amount of donation added to this tax year" -> By.cssSelector("#main-content > div > div > dl > div:nth-child(7) > dd.govuk-summary-list__actions.govuk-\\!-width-one-third > a"),
-    "change value of shares or securities" -> By.cssSelector("#main-content > div > div > dl > div:nth-child(8) > dd.govuk-summary-list__actions.govuk-\\!-width-one-third > a"),
-    "change value of land or property" -> By.cssSelector("#main-content > div > div > dl > div:nth-child(9) > dd.govuk-summary-list__actions.govuk-\\!-width-one-third > a"),
-    "change value of shares, securities, land or property" -> By.cssSelector("#main-content > div > div > dl > div:nth-child(10) > dd.govuk-summary-list__actions.govuk-\\!-width-one-third > a"),
-    "change overseas charities shares, securities, land or property were donated to" -> By.cssSelector("#main-content > div > div > dl > div:nth-child(11) > dd.govuk-summary-list__actions.govuk-\\!-width-one-third > a"),
-    "change charity name" -> By.cssSelector("#main-content > div > div > div.govuk-form-group > dl > div > dd > ul > li:nth-child(1) > a"),
-
-    //Crystallisation Elements
-    "I agree - Submit Income Tax Return" -> By.cssSelector("#agree"),
-    "Client SA UTR"                      -> By.id("utr"),
-    "View and Change Continue"           -> By.id("continue-button"),
-
-    //Construction Industry Scheme Elements
-    "cis deductions"             -> By.id("cis_link"),
-    "first contractor"           -> By.cssSelector("tbody > tr.govuk-table__row:nth-child(1) > th > a"),
-    "second contractor"          -> By.cssSelector("tbody > tr.govuk-table__row:nth-child(2) > th > a"),
-    "third contractor"           -> By.cssSelector("tbody > tr.govuk-table__row:nth-child(3) > th > a"),
-    "cis return to overview"     -> By.id("return-to-overview-button-id"),
-    "first month view"           -> By.cssSelector("#main-content > div > div > dl > div > dd > a"),
-    "second month view"          -> By.cssSelector("#main-content > div > div > dl > div:nth-child(2) > dd > a"),
-    "return to cis summary"      -> By.id("return-to-summary-button-id"),
-    "return to contractor"       -> By.id("return-to-contractor-button-id"),
-    "first month change"         -> By.cssSelector("#main-content > div > div > dl > div:nth-child(1) > dd:nth-child(2) > a"),
-    "first month remove"         -> By.cssSelector("#main-content > div > div > dl > div:nth-child(1) > dd:nth-child(3) > a"),
-    "change end of tax month"    -> By.cssSelector("#main-content > div > div > dl > div:nth-child(2) > dd:nth-child(3) > a"),
-    "change labour"              -> By.cssSelector("#main-content > div > div > dl > div:nth-child(3) > dd:nth-child(3) > a"),
-    "change cis deduction"       -> By.cssSelector("#main-content > div > div > dl > div:nth-child(4) > dd:nth-child(3) > a"),
-    "change paid for materials"  -> By.cssSelector("#main-content > div > div > dl > div:nth-child(5) > dd:nth-child(3) > a"),
-    "change cost of materials"   -> By.cssSelector("#main-content > div > div > dl > div:nth-child(6) > dd:nth-child(3) > a"),
-    "Add another contractor"     -> By.id("add-contractor"),
-    "Contractor name"            -> By.id("contractorName"),
-    "Employer Reference Number"  -> By.id("employerReferenceNumber"),
-    "Add another CIS deduction"  -> By.id("add-another-link"),
-    "Add a CIS deduction"        -> By.id("add-link"),
-    "Tax month ending"           -> By.id("month"),
-    "Add a contractor"           -> By.id("add-contractor"),
-    "CIS save and continue"      -> By.id("save-and-continue-button-id"),
-    "Remove period"              -> By.id("remove-period-button-id"),
-    "CIS deductions change"      -> By.xpath("//*[@id=\"main-content\"]/div/div/dl/div/dd[2]/a/span[1]"), //("//dt[contains(text,'')]/../dd[2]/a/span[1]"),
-    "CIS You cannot update text" -> By.cssSelector("#main-content > div > div > div.govuk-inset-text"),
-    "Updated for Cis"            -> By.xpath("//*[contains(text(),'Updated')]/../../span/a[@id='cis_link']"),
-    "Cannot Update for Cis"      -> By.xpath("//*[contains(text(),'Cannot update')]/../../span/a[@id='cis_link']"),
-
-    //State Benefits Elements
-    // "State benefits"                          -> By.id("stateBenefits_link"),
-    "State benefits"                          -> By.cssSelector("#stateBenefits_link"),
-    "Jobseeker’s Allowance"                   -> By.cssSelector("#main-content > div > div > ul > li:nth-child(2) > div.govuk-task-list__name-and-hint > a"),
-    "Add missing claim"                       -> By.id("add-missing-claim-button-id"),
-    "claim day"                               -> By.id("value-for-day"),
-    "claim month"                             -> By.id("value-for-month"),
-    "claim year"                              -> By.id("value-for-year"),
-    "remove claim"                            -> By.id("remove-link-id"),
-    "Remove claim"                            -> By.id("remove-claim-button-id"),
-    "Restore claim"                           -> By.id("restore-claim-button-id"),
-    "first view"                              -> By.cssSelector("#benefits-summary-list-id > div:nth-child(1) > dd.govuk-summary-list__actions > a > span:nth-child(1)"),
-    "first ignored view"                      -> By.cssSelector("#ignored-benefits-summary-list-id > div:nth-child(1) > dd:nth-child(3) > a:nth-child(1)"),
-    "first view in year"                      -> By.cssSelector("div.govuk-summary-list__row:nth-child(1) > dd:nth-child(2) > a:nth-child(1)"),
-    "state benefit change start date"         -> By.cssSelector("div.govuk-summary-list__row:nth-child(1) > dd:nth-child(3) > a:nth-child(1) > span:nth-child(1)"),
-    "state benefit change end date question"  -> By.cssSelector("div.govuk-summary-list__row:nth-child(2) > dd:nth-child(3) > a:nth-child(1) > span:nth-child(1)"),
-    "state benefit change end date"           -> By.cssSelector("div.govuk-summary-list__row:nth-child(3) > dd:nth-child(3) > a:nth-child(1) > span:nth-child(1)"),
-    "state benefit change tax taken question" -> By.cssSelector("div.govuk-summary-list__row:nth-child(4) > dd:nth-child(3) > a:nth-child(1) > span:nth-child(1)"),
-    "state benefit change amount"             -> By.cssSelector("div.govuk-summary-list__row:nth-child(5) > dd:nth-child(3) > a:nth-child(1) > span:nth-child(1)"),
-    "state benefit save and continue"         -> By.id("save-and-continue-button-id"),
-    "state benefit back"                      -> By.id("back-link"),
-    "Employment and Support Allowance"        -> By.cssSelector("#main-content > div > div > ul > li:nth-child(1) > div.govuk-task-list__name-and-hint > a"),
-    "The date you started getting Jobseeker’s Allowance must include a day" -> By.id("value-for-error"),
-    "The date you started getting Jobseeker’s Allowance must include a month"   -> By.id("emptyMonth-error"),
-    "The date you started getting Jobseeker’s Allowance must include a year" -> By.id("emptyYear-error"),
-    "Yes, claim Jobseeker's Allowance" -> By.id("value"),
-    "Yes, claim Employment and Support Allowance" -> By.id("value"),
-    "continue, missing claim" -> By.id("add-missing-claim-button-id"),
-
-    //View Estimation Elements
-    "Update tax calculation"   -> By.id("updateTaxCalculation"),
-    "Go to Income Tax Account" -> By.id("continue-button"),
-
-    //Pensions Elements
-    "pensions" -> By.id("pensions_link"),
-    "Payments into pensions" -> By.id("payments-into-pensions-link"),
-    "Unauthorised payments from pensions" -> By.id("unauthorised-payments-from-pensions-link"),
-    "Yes, unauthorised payments that resulted in a surcharge" -> By.id("unauthorisedPayments"),
-    "Yes, unauthorised payments that did not result in a surcharge" -> By.id("unauthorisedPayments-2"),
-    "No unauthorised payments" -> By.id("unauthorisedPayments-4"),
-    "amount in pounds" -> By.id("amount-2"),
-    "PSTR" -> By.id("taxReferenceId"),
-    "yes on" -> By.id("value-yes"),
-    "Add another Pensions Scheme Tax Reference" -> By.id("add-another-pension-link"),
-    "Change PSTR" -> By.cssSelector("#pensionIncomeSummaryList > dl > div:nth-child(1) > dd.hmrc-add-to-a-list__change > a"),
-    "Remove PSTR" -> By.cssSelector("#pensionIncomeSummaryList > dl > div:nth-child(1) > dd.hmrc-add-to-a-list__remove > a"),
-    "Remove Reference" -> By.id("remove-pension-button-id"),
-    "Overseas pensions" -> By.id("overseas-pensions-summary-link"),
-    "Income from overseas pensions" -> By.id("income-from-overseas-pensions-link"),
-    "Country field" -> By.id("countryId"),
-    "Country List" -> By.id("countryId__listbox"),
-    "amount before tax" -> By.id("amount-1"),
-    "tax paid" -> By.id("amount-2"),
-    "amount in pounds" -> By.id("amount-2"),
-    "pensions scheme save and continue" -> By.id("continue"),
-    "add another pension scheme" -> By.id("add-another-pension-link"),
-    "Name of pension scheme" -> By.id("providerName"),
-    "Pension Scheme Tax Reference" -> By.id("schemeReference"),
-    "Pension provider address" -> By.id("providerAddress"),
-    "Remove pension scheme" -> By.cssSelector("#pensionTransferSummaryList > dl > div > dd.hmrc-add-to-a-list__remove > a > span:nth-child(1)"),
-    "Remove Scheme" -> By.id("remove-pension-button-id"),
-    "Add a scheme" -> By.id("AddAScheme"),
-    "Change scheme" -> By.cssSelector("#pensionTransferSummaryList > dl > div > dd.hmrc-add-to-a-list__change > a > span:nth-child(1)"),
-    "Overseas transfer charges" -> By.id("transfer-into-overseas-pensions-link"),
-    "Pension lifetime allowance" -> By.id("pension-lifetime-allowance-link"),
-    "Adding another Pensions Scheme Tax Reference" -> By.id("add-another-link"),
-    "Pension annual allowance" -> By.id("annual-allowances-link"),
-    "Money purchase annual allowance" -> By.id("reducedAnnualAllowanceType"),
-    "Tapered annual allowance" -> By.id("reducedAnnualAllowanceType-2"),
-    "Remove allowance scheme"  -> By.cssSelector("div > dl > div:nth-child(1) > dd.hmrc-add-to-a-list__remove > a"),
-    "Change allowance scheme" -> By.cssSelector("div > dl > div:nth-child(1) > dd.hmrc-add-to-a-list__change > a"),
-    "Short service refunds" -> By.id("short-service-refunds-link"),
-    "Income from pensions" -> By.id("income-from-pensions-summary-link"),
-    "State pension" -> By.id("state-pension-link"),
-    "pension day" -> By.id("day"),
-    "pension month" -> By.id("month"),
-    "pension year" -> By.id("year"),
-    "Other UK Pensions" -> By.id("uk-pension-income-link"),
-    "Pension Identification" -> By.id("pensionId"),
-    "Payment into overseas pensions" -> By.id("payments-into-overseas-pensions-link"),
-    "Customer reference number" -> By.id("pensionsCustomerReferenceNumberId"),
-    "Migrant member relief" -> By.id("mmrValue") ,
-    "Double taxation relief" -> By.id("dtrValue") ,
-    "Transitional corresponding relief" -> By.id("tcrValue") ,
-    "None of these" -> By.id("noTaxRelief") ,
-    "SF74 reference" -> By.id("sf74ReferenceId"),
-    "QOPS Number" -> By.id("qopsReferenceId"),
-    "QOPS Reference" -> By.id("schemeReference"),
-
-    //Gains Elements
-    "Gains from policies and contracts"                               -> By.id("gains_link"),
-    "Life Insurance"                                                  -> By.cssSelector("[value='lifeInsurance']"),
-    "Full or part surrender"                                          -> By.cssSelector("[value='Full or part surrender']"),
-    "Other"                                                           -> By.id("other-text"),
-    "Sale or assignment of a policy"                                  -> By.id("policy-event-3"),
-    "What caused this gain?"                                          -> By.id("other-text-input"),
-    "Policy number"                                                   -> By.id("value"),
-    "Gains Year"                                                      -> By.id("year"),
-    "Policy remove"                                                   -> By.id("policy-remove"),
-    "remove"                                                          -> By.id("remove"),
-    "Updated Tag"                                                     -> By.xpath("//*[contains(text(),'Updated')]/../../span/a[@id='gains_link']"),
-    "Add a policy"                                                    -> By.id("continue"),
-    "Excluded for Gains"                                              -> By.xpath("//*[contains(text(),'Excluded')]/../../span/a[@id='gains_link']"),
-    "Voided ISA"                                                      -> By.cssSelector("[value='voidedIsa']"),
-    "Add another gain from a different policy"                        -> By.xpath("//*[contains(text(),'Add another gain from a different policy')]"),
-    "Gains status"                                                    -> By.xpath("//*[@id='gains_link']/parent::span/parent::li/span[2]/strong"),
-    "Chane Policy"                                                    -> By.id("policy-change"),
-
-    //About You
-    "About you"                                                   -> By.cssSelector("#main-content > div > div > ul > li:nth-child(1) > div.govuk-task-list__name-and-hint > a"),
-    "About you tag"                                               -> By.cssSelector("#task-list-1-status"),
-    "I was not resident in the UK"                                -> By.cssSelector("[value='WasNot']"),
-    "I was a dual resident in the UK and another country"         -> By.cssSelector("[value='Was']"),
-    "No, my client did not donate to charity"                     -> By.id("CharitableDonations-6"),
-    "My income"                                                   -> By.cssSelector("[value='selfIncome']"),
-    "Tax avoidance"                                               -> By.cssSelector("[value='taxAvoidance']"),
-    "No, my client was not in any tax avoidance schemes"          -> By.id("TaxAvoidance-4"),
-    "My partner’s income"                                 -> By.cssSelector("[value='MyPartnersIncome']"),
-    "I do not have a partner"                             -> By.cssSelector("[value='NoPartner']"),
-    "Donations using Gift Aid"                            -> By.id("CharitableDonations"),
-    "Gifts of shares or securities"                       -> By.id("CharitableDonations-2"),
-    "Gifts of land or property"                           -> By.id("CharitableDonations-3"),
-    "Gifts of overseas charities"                           -> By.id("CharitableDonations-4"),
-    "No, I did not donate to charity"                     -> By.id("CharitableDonations-5"),
-    "yes, but permanent home is abroad"                   -> By.cssSelector("[value='domiciled']"),
-    "Not a resident"                                      -> By.cssSelector("[value='nonUK']"),
-    "My client was not resident in the UK"                -> By.cssSelector("[value='nonResident']"),
-    "No residency"                                        -> By.cssSelector("[value='nonUK']"),
-
-    //Income from work and taxable state benefits
-    "Income from work and taxable state benefits"      -> By.cssSelector("#main-content > div > div > ul > li:nth-child(2) > div.govuk-task-list__name-and-hint > a"),
-    "Income from work and taxable state benefits tag"  -> By.cssSelector("#task-list-2-status"),
-    "Employed"                                         -> By.cssSelector("[value='employed']"),
-    "Self-employed"                                    -> By.cssSelector("[value='selfEmployed']"),
-    "Jobseeker's Allowance"                            -> By.id("JobseekersAllowance"),
-    "Contribution-based Employment and Support Allowance"                            -> By.id("JobseekersAllowance-2"),
-    "Member of Parliament"                             -> By.cssSelector("[value='MemberOfParliament']"),
-    "None of the above"                                -> By.id("AboutYourWork-4"),
-    "No, I did not get either of these state benefits" -> By.id("JobseekersAllowance-4"),
-    "Contribution-based"                               -> By.id("JobseekersAllowance-2"),
-    "Partnership"                                      -> By.cssSelector("[value='Partnership']"),
-    "Hidden message"                                   -> By.cssSelector("[class='govuk-body govuk-!-padding-2']"),
-    "yes, but permanent home is abroad"                -> By.id("value_1"),
-    "checkboxes"                                       ->By.cssSelector("[class='govuk-checkboxes__input']"),
-
-    // Property UK
-    "Property"                 -> By.id("property_link"),
-    "UK property" -> By.cssSelector("#main-content > div > div > h2"),
-    "UK property About" -> By.id("property_about_link"),
-    "UK Property details" -> By.cssSelector("#main-content > div > div > h1"),
-    "property continue" -> By.cssSelector("#main-content > div > div > form > button"),
-    "How much total income did you get from your UK properties?" -> By.cssSelector("#main-content > div > div > form > div > fieldset > legend > h1"),
-    "first property" -> By.id("value_0"),
-    "second property" -> By.id("value_1"),
-    "How much total income did your client get from all of their properties?" -> By.cssSelector("#main-content > div > div > form > div > fieldset > legend > h1"),
-    "Which rental property do you have in the UK?" -> By.cssSelector("#main-content > div > div > form > h1"),
-    "Which rental property does your client have in the UK?" -> By.cssSelector("#main-content > div > div > form > h1"),
-    "Property Rentals" -> By.id("value_0"),
-    "Have you finished this section?" -> By.cssSelector("#main-content > div > div > form > div > fieldset > legend > h1"),
-    "HYFTS AboutPropertyYes" -> By.id("isAboutPropertyRentalsSectionFinished"),
-    "HYFTS AboutPropertyNo" -> By.id("isAboutPropertyRentalsSectionFinished-no"),
-    "HYFTS,Yes" -> By.id("isAboutPropertyComplete"),
-    "HYFTS,No" -> By.id("aboutPropertyCompleteYesOrNo-no"),
-    "Yes HYFTS" -> By.id("isAboutPropertyRentalsSectionFinished"),
-    "No HYFTS" -> By.id("isAboutPropertyRentalsSectionFinished-no"),
-    "IncomeYes,HYFTS" -> By.id("isIncomeSectionFinished"),
-    "IncomeNo,HYFTS" -> By.id("isIncomeSectionFinished-no"),
-    "AdjustmentsYes,HYFTS" -> By.id("isRentalsAdjustmentsComplete"),
-    "AdjustmentsNo,HYFTS" -> By.id("isRentalsAdjustmentsComplete-no"),
-    "ExpensesYes,HYFTS" -> By.id("isExpensesSectionFinished"),
-    "ExpensesNo,HYFTS" -> By.id("isExpensesSectionFinished-no"),
-    "AllowancesYes,HYFTS" -> By.id("isAllowancesSectionFinished"),
-    "AllowancesNo,HYFTS" -> By.id("isAllowancesSectionFinished-no"),
-    "sbaYes,HYFTS" -> By.id("isSbaSectionFinished"),
-    "sbaNo,HYFTS" -> By.id("isSbaSectionFinished-no"),
-    "esbaYes,HYFTS" -> By.id("isEsbaSectionFinished"),
-    "esbaNo,HYFTS" -> By.id("isEsbaSectionFinished-no"),
-    "UK Property continue" -> By.cssSelector("#main-content > div > div > form > button"),
-    "Check your answers" -> By.cssSelector("#main-content > div > div > h1"),
-    "Check your answers continue" -> By.id("continue"),
-    "Property rentals about" -> By.id("rentals_about_link"),
-    "Property rentals" -> By.cssSelector("#main-content > div > div > h1"),
-    "Property rentals" -> By.id("value_0"),
-    "Property rentals continue" -> By.id("continue"),
-    "Check Your Answers" -> By.cssSelector("#main-content > div > div > h1"),
-    "Check Your Answers continue" -> By.id("continue"),
-    "UK property rentals about" -> By.id ("rentals_about_link"),
-    "UK property rentals" -> By.cssSelector("#main-content > div > div > h2:nth-child(4)"),
-    "Property rentals Yes" -> By.id("value"),
-    "Property rentals No"-> By.id("value-no"),
-    "Do you want to claim property income allowance?" -> By.cssSelector("#main-content > div > div > form > div > fieldset > legend > h1"),
-    "Does your client want to claim property income allowance?" -> By.cssSelector("#main-content > div > div > form > div > fieldset > legend"),
-    "Yes, claim property income allowance" -> By.id("isClaimPropertyIncomeAllowance"),
-    "No, claim expenses" -> By.id("isClaimPropertyIncomeAllowance-no"),
-    "Check Your Answers"-> By.cssSelector("#main-content > div > div > h1"),
-    "Are your total expenses less than £1,000?" -> By.cssSelector("#main-content > div > div > dl > div:nth-child(1) > dt"),
-    "Would you like to claim property income allowance?" -> By.cssSelector("#main-content > div > div > dl > div:nth-child(2) > dt"),
-    "Change" -> By.cssSelector("#main-content > div > div > form > dl > div > dd.govuk-summary-list__actions > a"),
-    "Rentals Income" -> By.id("rentals_income_link"),
-    "Property Income" -> By.cssSelector("#main-content > div > div > h1"),
-    "Non-UK resident landlord" -> By.cssSelector("<h1 class=\"govuk-heading-l\">Non-UK resident landlord</h1>"),
-    "Deducting Tax from non-UK landlords" -> By.cssSelector("<h1 class=\"govuk-heading-l\">Deducting Tax from non-UK landlords</h1>"),
-    "Tax deducted Yes" -> By.id("deductingTax"),
-    "Tax deducted No" -> By.id("deductingTax-no"),
-    "Tax Deducted Amount" -> By.id("taxDeductedAmount"),
-    "Property rental income" -> By.cssSelector("#main-content > div > div > form > h1"),
-    "How much total income did you recieve from from all of your property rentals?" -> By.cssSelector("#main-content > div > div > form > h3"),
-    "Property rental amount" -> By.id("propertyRentalIncome"),
-    "Premiums for the grant of a lease" -> By.cssSelector("<h1 class=\"govuk-heading-l\">Premiums for the grant of a lease</h1>"),
-    "Have you calculated the figure yourself?" -> By.cssSelector("<h1 class=\"govuk-heading-l\">Have you calculated the figure yourself?</h1>"),
-    "Has your client calculated the premium for the grant of a lease taxable amount themselves?" -> By.cssSelector("<h1 class=\"govuk-heading-l\">Has your client calculated the premium for the grant of a lease taxable amount themselves?</h1>"),
-    "Yes,calculated amount" -> By.id("calculatedFigureYourselfAmount"),
-    "Yes,calculated" -> By.id("calculatedFigureYourself"),
-    "No,calculated" -> By.id("calculatedFigureYourself-no"),
-    "How much did you receive for a grant of a lease?" -> By.cssSelector("#main-content > div > div > form > div > h1 > label"),
-    "How much did your client receive for the grant of a lease?" -> By.cssSelector("#main-content > div > div > form > div > h1 > label"),
-    "grant lease amount" -> By.id("receivedGrantLeaseAmount"),
-    "How many complete 12 month periods were in the term of the lease?" -> By.cssSelector("#main-content > div > div > form > div > h1 > label"),
-    "month period" -> By.id("yearLeaseAmount"),
-    "Premiums for a grant of a lease" -> By.cssSelector("#main-content > div > div > h1"),
-    "Yes, premiums for a grant of lease" -> By.id("isPremiumForLease"),
-    "No, premiums for a grant of lease" -> By.id("isPremiumForLease-no"),
-    "Yes, premiums for the grant of lease" -> By.id("premiumsGrantLease"),
-    "No, premiums for the grant of lease" -> By.id("premiumsGrantLease-no"),
-    "Amount, premiums for a grant of lease" -> By.id("premiumsGrantLeaseAmount"),
-    "premium amount"-> By.id("PremiumPayment"),
-    "Reverse premiums received" -> By.cssSelector("<h1 class=\"govuk-heading-l\">Reverse premiums received</h1>"),
-
-    "premium recieved amount"-> By.id("reversePremiums"),
-    "Yes,premiums recieved" -> By.id("reversePremiumsReceived"),
-    "No,premiums recieved" -> By.id("reversePremiumsReceived-no"),
-    "Other income from property" -> By.cssSelector("<h1 class=\"govuk-heading-l\">Other income from property</h1>"),
-    "Other income amount" -> By.id("otherIncomeFromProperty"),
-    "Check Your Answers" -> By.cssSelector("#main-content > div > div > h1"),
-    "Rentals Adjustments" -> By.id("rentals_adjustments_link"),
-    "How much was your private use adjustment?" -> By.cssSelector("<h1 class=\"govuk-heading-l\">How much was your private use adjustment?</h1>"),
-    "Private use adjustment Amount"-> By.id("privateUseAdjustmentAmount"),
-    "Balancing charge" -> By.cssSelector("<h1 class=\"govuk-heading-l\">Balancing charge</h1>"),
-    "BalancingChargeYes" -> By.id("isBalancingCharge"),
-    "BalancingChargeNo" -> By.id("isBalancingCharge-no"),
-    "How much is the balancing charge? Amount" -> By.id("balancingChargeAmount"),
-    "Your Property Income Allowance claim" -> By.cssSelector("#main-content > div > div > h1"),
-    "Enter your share of PIA for this property Amount" -> By.id("propertyIncomeAllowance"),
-    "Business premises renovation allowance balancing charge" -> By.cssSelector("<h1 class=\"govuk-heading-l\">Business premises renovation allowance balancing charge</h1>"),
-    "Business premises renovation allowance balancing charge Yes" -> By.id("renovationAllowanceBalancingChargeForm"),
-    "Business premises renovation allowance balancing charge Amount" -> By.id("renovationAllowanceBalancingChargeAmount"),
-
-    "Yes,How much is the renovation allowance balancing charge?" -> By.id("renovationAllowanceBalancingCharge"),
-    "How much is the renovation allowance balancing charge? amount" -> By.id("renovationAllowanceBalancingChargeAmount"),
-    "No,How much is the balancing charge?" -> By.id("renovationAllowanceBalancingCharge-no"),
-    "Residential finance costs" -> By.cssSelector("<h1 class=\"govuk-heading-l\">Residential finance costs</h1>"),
-    "How much were your residential finance costs? Amount" -> By.id("residentialFinanceCost"),
-    "Unused residential costs brought forward" -> By.cssSelector("<h1 class=\"govuk-heading-l\">Unused residential costs brought forward</h1>"),
-    "How much unused residential costs do you want to bring forward? Amount" -> By.id("unusedResidentialFinanceCost"),
-    "Did you have any unused losses for your property rental in previous years?" -> By.cssSelector("<h1 class=\"govuk-heading-l\">Did you have any unused losses for your property rental in previous years?</h1>"),
-    "Yes, did you have any unused losses for your property rental in previous years?" -> By.id("isUnusedLossesBroughtForward"),
-    "Yes, how much unused residential costs do you want to bring forward? Amount" -> By.id("unusedLossesBroughtForwardAmount"),
-    "When you reported the loss" -> By.xpath("//*[@id=\"main-content\"]/div/div/form/h1"),
-    "2020 to 2021" -> By.id("value_2"),
-    "Check Your Answers" -> By.cssSelector("#main-content > div > div > h1"),
-    "Rentals Expenses" -> By.id("rentals_expenses_link"),
-    "Consolidated or individual expenses" -> By.cssSelector("#main-content > div > div > h1"),
-    "Yes,consolidated expenses" -> By.id("isConsolidatedExpenses"),
-    "No,consolidated expenses" -> By.id("isConsolidatedExpenses-no"),
-    "consolidated expenses amount" -> By.id("consolidatedExpensesAmount"),
-    "Check Your Answers" -> By.cssSelector("#main-content > div > div > h1"),
-    "Rents, rates and insurance costs for your property" -> By.cssSelector("<h1 class=\"govuk-heading-l\">Rents, rates and insurance costs for your property</h1>"),
-    "Rents, rates and insurance amount" -> By.id("rentsRatesAndInsurance"),
-    "Property repairs and maintenance" -> By.cssSelector("<h1 class=\"govuk-heading-l\">Property repairs and maintenance</h1>"),
-    "Property repairs and maintenance amount" -> By.id("repairsAndMaintenanceCosts"),
-    "Loan interest or other financial costs" -> By.cssSelector("<h1 class=\"govuk-heading-l\">Loan interest or other financial costs</h1>"),
-    "Loan interest or other financial costs amount" -> By.id("loanInterestOrOtherFinancialCost"),
-    "Legal, management or other professional fees" -> By.cssSelector("<h1 class=\"govuk-heading-l\">Legal, management or other professional fees</h1>"),
-    "Legal, management or other professional fees amount" -> By.id("otherProfessionalFees"),
-    "Costs of services provided, including wages" -> By.cssSelector("<h1 class=\"govuk-heading-l\">Costs of services provided, including wages</h1>"),
-    "Costs of services provided, including wages amount" -> By.id("costsOfServicesProvided"),
-    "Property business travel costs" -> By.cssSelector("<h1 class=\"govuk-heading-l\">Property business travel costs</h1>"),
-    "Property business travel costs amount" -> By.id("propertyBusinessTravelCosts"),
-    "Other allowable property expenses" -> By.cssSelector("<h1 class=\"govuk-heading-l\">Other allowable property expenses</h1>"),
-    "Other allowable property expenses amount" -> By.id("otherAllowablePropertyExpenses"),
-    "Rentals Allowances" -> By.id("rentals_allowances_link"),
-    "Capital allowances for a car" -> By.cssSelector("<h1 class=\"govuk-heading-l\">Capital allowances for a car</h1>"),
-    "Yes,capital allowances for a car" -> By.id("isCapitalAllowancesForACar"),
-    "No,capital allowances for a car" -> By.id("isCapitalAllowancesForACar-no"),
-    "Capital allowances for a car amount" -> By.id("capitalAllowancesForACarAmount"),
-    "Check Your Answers" -> By.cssSelector("#main-content > div > div > h1"),
-    "Annual investment allowance" -> By.cssSelector("#main-content > div > div > h1"),
-    "Annual investment allowance amount" -> By.id("annualInvestmentAllowanceAmount"),
-    "Electric charge-point allowance" -> By.cssSelector("#main-content > div > div > form > h1"),
-    "Yes,electric charge point allowance" -> By.id("electricChargePointAllowanceYesNo"),
-    "No,electric charge point allowance" -> By.id("electricChargePointAllowanceYesNo-no"),
-    "Electric charge-point allowance amount" -> By.id("electricChargePointAllowanceAmount"),
-    "Zero-emission car allowance" -> By.cssSelector("#main-content > div > div > h1"),
-    "Zero-emission car allowance amount" -> By.id("zeroEmissionCarAllowanceAmount"),
-    "Zero-emission goods vehicle allowance" -> By.cssSelector(" #main-content > div > div > h1"),
-    "Zero-emission goods vehicle allowance amount" -> By.id("zeroEmissionGoodsVehicleAllowanceAmount"),
-    "Business premises renovation allowance" -> By.cssSelector("#main-content > div > div > h1"),
-    "Business premises renovation allowance amount" -> By.id("businessPremisesRenovationAmount"),
-    "Replacement of domestic goods" -> By.cssSelector("#main-content > div > div > h1"),
-    "Replacement of domestic goods amount" -> By.id("replacementOfDomesticGoodsAmount"),
-    "Other capital allowances" -> By.cssSelector("#main-content > div > div > h1"),
-    "Other capital allowances amount" -> By.id("otherCapitalAllowanceAmount"),
-    "Rentals Structures and Buildings Allowance" -> By.id("rentals_structures_and_building_allowance_link"),
-    "Do you want to claim structures and buildings allowance?" -> By.cssSelector("#main-content > div > div > form > h1"),
-    "Yes, claimStructureBuildingAllowance" -> By.id("claimStructureBuildingAllowance"),
-    "No,claimStructureBuildingAllowance" -> By.id("claimStructureBuildingAllowance-no"),
-    "Structures Buildings Allowance" -> By.cssSelector("#main-content > div > div > h1"),
-    "Add Claim" -> By.id("continue"),
-    "Qualifying date" -> By.cssSelector("#main-content > div > div > form > h1"),
-    "sba Day" -> By.id("structureBuildingQualifyingDate.day"),
-    "sba Month" -> By.id("structureBuildingQualifyingDate.month"),
-    "sba Year" -> By.id("structureBuildingQualifyingDate.year"),
-    "Structures and buildings allowance qualifying amount" -> By.cssSelector("#main-content > div > div > form > h1"),
-    "Qualifying Amount" -> By.id("structureBuildingQualifyingAmount"),
-    "Qualifying amt" -> By.id("esbaQualifyingAmount"),
-    "How much is your structures and buildings allowance claim?" -> By.cssSelector("#main-content > div > div > form > h1"),
-    "How much is your structures and buildings allowance claim? amount" -> By.id("structureBuildingAllowanceClaim"),
-    "Structures and buildings allowance address" -> By.cssSelector("#main-content > div > div > h1"),
-    "Building name" -> By.id("buildingName"),
-    "Building number" -> By.id("buildingNumber"),
-    "Postcode" -> By.id("postcode"),
-    "Check Your Answers" -> By.cssSelector("#main-content > div > div > h1"),
-    "Save and conitnue" -> By.id("continue"),
-    "Your structures and buildings allowance claims" -> By.cssSelector("#main-content > div > div > form > h1"),
-    "Do you want to add another claim?" -> By.cssSelector("#main-content > div > div > form > div > fieldset > legend > h1"),
-    "Yes,anotherClaim" -> By.id("anotherClaim"),
-    "No,anotherClaim" -> By.id("anotherClaim-no"),
-    "RemoveClaim" -> By.cssSelector("#main-content > div > div > form > dl > div:nth-child(1) > dd.govuk-summary-list__actions.w-25 > ul > li:nth-child(2) > a"),
-    "RemoveYes" -> By.id("sbaRemoveConfirmation"),
-    "RemoveNo" -> By.id("sbaRemoveConfirmation-no"),
-    "Do you want to remove this structures and buildings allowance £1,659 claim?" -> By.cssSelector("#main-content > div > div > form > div > fieldset > legend > h1"),
-    "Rentals Enhanced Structures and Buildings Allowance" -> By.id("rentals_enhanced_structures_and_building_allowance_link"),
-    "Do you want to claim enhanced structures and buildings allowance?" -> By.cssSelector("#main-content > div > div > form > div > fieldset > legend"),
-    "Yes, claim" -> By.id("claimEnhancedStructureBuildingAllowance"),
-    "No, claim" -> By.id("claimEnhancedStructureBuildingAllowance-no"),
-    "esba Day" -> By.id("esbaQualifyingDate.day"),
-    "esba Month" -> By.id("esbaQualifyingDate.month"),
-    "esba Year" -> By.id("esbaQualifyingDate.year"),
-    "Enhanced structures and buildings allowance claims" -> By.cssSelector("#main-content > div > div > h1"),
-    "How much is your enhanced structures and buildings allowance claim?" -> By.cssSelector("#main-content > div > div > form > div > h1 > label"),
-    "How much is your enhanced structures and buildings allowance claim? amount" -> By.id("esbaClaim"),
-    "Enhanced structures and buildings allowance address" -> By.cssSelector("#main-content > div > div > h1"),
-    "Your enhanced structures and buildings allowance claims" -> By.cssSelector("#main-content > div > div > form > h1"),
-
-    // Foreign Property
-    "Foreign property select country" -> By.cssSelector("a[aria-describedby='foreign_property_select_country']"),
-    "Country" -> By.id("country-autocomplete"),
-    "country" -> By.id("country"),
-    "Country option" -> By.id("country-autocomplete__option--0"),
-
-    "Foreign property details" -> By.cssSelector("#main-content > div > div > form > div > fieldset > legend > h1"),
-    "Which country did you receive this income from?" -> By.cssSelector("#main-content > div > div > h1"),
-    "Which country did your client receive this income from?" -> By.cssSelector("#main-content > div > div > h1"),
-    "Countries where you rented out property" -> By.cssSelector("#main-content > div > div > h1"),
-    "Countries where your client rented out property" -> By.cssSelector("#main-content > div > div > h1"),
-    "Foreign property Declare expenses" -> By.id("claimPropertyIncomeAllowanceOrExpenses-no"),
-    "Foreign property Property income allowance" -> By.id("claimPropertyIncomeAllowanceOrExpenses"),
-    "Use the property income allowance" -> By.id("claimPropertyIncomeAllowanceOrExpenses"),
-    "How much total income did your client get from all of their properties?" -> By.cssSelector("#main-content > div > div > form > div > fieldset > legend > h1"),
-    "How much total income did your client get from all of their foreign properties?" -> By.cssSelector("#main-content > div > div > form > div > fieldset > legend > h1"),
-
-    // Foreign tax
-    "Foreign tax Spain" -> By.id("foreign_property_income_tax_ESP"),
-    "Foreign tax Italy" -> By.id("foreign_property_income_tax_ITA"),
-    "Did you pay foreign tax on the income from your property in Spain?" -> By.cssSelector("#main-content > div > div > h1"),
-    "Did your client pay foreign tax on the income from their property in Italy?" -> By.cssSelector("<h1 class=\"govuk-heading-l\">Did your client pay foreign tax on the income from their property in Italy?</h1>"),
-    "Foreign income tax Yes" -> By.id("foreignIncomeTaxForm"),
-    "Foreign income tax Amount" -> By.id("foreignTaxPaidOrDeducted"),
-    "Do you want to claim Foreign Tax Credit Relief?" -> By.cssSelector("<h1 class=\"govuk-heading-l\">Do you want to claim Foreign Tax Credit Relief?</h1>"),
-    "Does your client want to claim Foreign Tax Credit Relief?" -> By.cssSelector("<h1 class=\"govuk-heading-l\">Does your client want to claim Foreign Tax Credit Relief?</h1>"),
-
-    // Foreign income
-    "Foreign income Spain" -> By.id("foreign_property_income_ESP"),
-    "Foreign income Italy" -> By.id("foreign_property_income_ITA"),
-    "Income from your property in Spain" -> By.cssSelector("<h1 class=\"govuk-heading-l\">Income from your property in Spain</h1>"),
-    "Income from your client’s property in Italy" -> By.cssSelector("<h1 class=\"govuk-heading-l\">Income from your client’s property in Italy</h1>"),
-    "Foreign property rental income" -> By.cssSelector("<h1 class=\"govuk-heading-l\">Foreign property rental income</h1>"),
-    "Rental income" -> By.id("rentIncome"),
-    "No, calculate it for me" -> By.id("calculatedPremiumLeaseTaxable-no"),
-    "12 month period" -> By.id("twelveMonthPeriodsInLease"),
-    "Have you calculated the taxable amount for the premium for the grant of a lease?" -> By.cssSelector("<h1 class=\"govuk-heading-l\">Have you calculated the taxable amount for the premium for the grant of a lease?</h1>"),
-    "No disagree premiums grant lease" -> By.id("foreignPremiumsGrantLeaseForm-no"),
-    "Calculated premiums grant lease" -> By.id("premiumsOfLeaseGrant"),
-    "Foreign other income amount" -> By.id("otherPropertyIncome"),
-
-    // Foreign expenses
-    "Foreign expenses Spain" -> By.id("foreign_property_expenses_ESP"),
-    "Expenses for your property in Spain" -> By.cssSelector("#main-content > div > div > h1"),
-    "Claim individual expenses foreign" -> By.id("consolidatedOrIndividualExpenses-no"),
-    "Rents, rates and insurance" -> By.id("foreignRentsRatesAndInsurance"),
-    "Property repairs and maintenance" -> By.id("foreignPropertyRepairsAndMaintenanceAmount"),
-    "Non-residential property finance costs" -> By.id("foreignNonResidentialPropertyFinanceCostsAmount"),
-    "Legal, management or other professional fees" -> By.id("foreignProfessionalFeesAmount"),
-    "Costs of services provided, including wages" -> By.id("foreignCostsOfServicesProvided"),
-    "Other allowable property expenses" -> By.id("otherAllowablePropertyExpensesAmount"),
-
-    //Foreign allowances
-    "Foreign allowances Spain" -> By.id("foreign_property_allowances_ESP"),
-    "Allowances for your property in Spain" -> By.cssSelector("#main-content > div > div > h1"),
-    "Zero-emission car allowance" -> By.cssSelector("<h1 class=\"govuk-heading-l\">Zero-emission car allowance</h1>"),
-    "Foreign zero-emission car allowance amount" -> By.id("zeroEmissionsCarAllowance"),
-    "Zero-emission goods vehicles" -> By.cssSelector("<h1 class=\"govuk-heading-l\">Zero-emission goods vehicles</h1>"),
-    "Zero-emission goods vehicles amount" -> By.id("zeroEmissionsGoodsVehicleAllowance"),
-    "Replacement of domestic goods" -> By.cssSelector("<h1 class=\"govuk-heading-l\">Replacement of domestic goods</h1>"),
-    "Foreign replacement of domestic goods amount" -> By.id("costOfReplacingDomesticItems"),
-    "Other capital allowances" -> By.cssSelector("<h1 class=\"govuk-heading-l\">Other capital allowances</h1>"),
-    "Foreign other capital allowances amount" -> By.id("otherCapitalAllowance"),
-
-    //Foreign structures and building allowance
-    "Foreign structures and building allowance Spain" -> By.id("foreign_structure_and_building_allowance_ESP"),
-    "Do you want to claim structures and buildings allowance?" -> By.cssSelector("<h1 class=\"govuk-heading-l\">Structures and buildings allowance"),
-    "Structures Buildings Allowance" -> By.cssSelector("<h1 class=\"govuk-heading-l\">Structures and buildings allowance claims"),
-    "Add a claim" -> By.id("continue"),
-    "Structures and buildings allowance qualifying date" -> By.cssSelector("<h1 class=\"govuk-heading-l\">Structures and buildings allowance qualifying date"),
-    "Qualifying Day" -> By.id("foreignStructureBuildingQualifyingDate.day"),
-    "Qualifying Month" -> By.id("foreignStructureBuildingQualifyingDate.month"),
-    "Qualifying Year" -> By.id("foreignStructureBuildingQualifyingDate.year"),
-    "Structures and buildings allowance qualifying amount" -> By.cssSelector("<h1 class=\"govuk-heading-l\">Structures and buildings allowance qualifying amount"),
-    "Foreign sba qualifying amount" -> By.id("foreignStructureBuildingQualifyingAmount"),
-    "How much is your structures and buildings allowance claim?" -> By.cssSelector("<h1 class=\"govuk-heading-l\">How much is your structures and buildings allowance claim?"),
-    "Foreign sba claim amount" -> By.id("foreignStructureBuildingAllowanceClaim"),
-    "Structure and building allowance address?" -> By.cssSelector("<h1 class=\"govuk-heading-l\">Structure and building allowance address?"),
-    "Foreign sba save and continue" -> By.id("continue"),
-
-    //Foreign adjustment
-    "Foreign adjustments Spain" -> By.id("foreign_property_adjustments_ESP"),
-    "Foreign adjustments Italy" -> By.id("foreign_property_adjustments_ITA"),
-    "Adjustments for your property in Spain" -> By.cssSelector("<h1 class=\"govuk-heading-l\">Adjustments for your property in Spain"),
-    "Adjustments for your property in Italy" -> By.cssSelector("<h1 class=\"govuk-heading-l\">Adjustments for your property in Italy"),
-    "Private use adjustment" -> By.cssSelector("<h1 class=\"govuk-heading-l\">Private use adjustment</h1>"),
-    "Your Property Income Allowance claim" -> By.cssSelector("<h1 class=\"govuk-heading-l\">Your Property Income Allowance claim</h1>"),
-    "PIA claim amount" -> By.id("propertyIncomeAllowanceClaimAmount"),
-    "Residential property finance costs" -> By.cssSelector("<h1 class=\"govuk-heading-l\">Residential property finance costs</h1>"),
-    "Residential property finance costs amount" -> By.id("residentialFinanceCostsAmount"),
-    "Replacement of domestic goods" -> By.cssSelector("<h1 class=\"govuk-heading-l\">Replacement of domestic goods</h1>"),
-    "Unused residential property finance costs brought forward" -> By.cssSelector("<h1 class=\"govuk-heading-l\">Unused residential property finance costs brought forward</h1>"),
-    "Foreign unused residential finance cost amount" -> By.id("foreignUnusedResidentialFinanceCostAmount"),
-    "Did your client have any unused losses for their property rental in previous years?" -> By.cssSelector("<h1 class=\"govuk-heading-l\">Did your client have any unused losses for their property rental in previous years?</h1>"),
-    "Foreign unused losses in previous years amount" -> By.id("unusedLossesPreviousYearsAmount"),
-    "When you reported the loss" -> By.cssSelector("<h1 class=\"govuk-heading-l\">When you reported the loss</h1>"),
-    "2021 to 2022" -> By.id("value_3"),
-
-    // UK & Foreign Property
-    "UK & Foreign Property About" -> By.id("uk_and_foreign_property_about_link"),
-    "Foreign property About" -> By.cssSelector("#main-content > div > div > ul > li:nth-child(3) > div.govuk-task-list__name-and-hint > a"),
-    "I do not want to report my property income" -> By.id("reportPropertyIncome-no"),
-    "I do not want to report my client's property income" -> By.id("value-no"),
-    "Less than £1,000" -> By.id("value_0"),
-    "Do you want to report your property income?" -> By.cssSelector("#main-content > div > div > h1"),
-    "Yes, I want to report my property income" -> By.id("value_0"),
-    "Which type of rental property do you have in the UK?" -> By.cssSelector("#main-content > div > div > h1"),
-    "Which type of rental property does your client have in the UK?" -> By.cssSelector("#main-content > div > div > h1"),
-    "Foreign countries where you rented out property" -> By.cssSelector("#main-content > div > div > h1"),
-    "Foreign countries where your client rented out property" -> By.cssSelector("#main-content > div > div > h1"),
-    "Which country did your client receive their foreign property rental income from?" -> By.cssSelector("#main-content > div > div > h1"),
-    "Which country did you receive your foreign property rental income from?" -> By.cssSelector("#main-content > div > div > h1"),
-    "Do you want to report your client’s property income?" -> By.cssSelector("#main-content > div > div > h1"),
-    "How much total income did your client get from all of their UK and foreign properties?" -> By.cssSelector("#main-content > div > div > form > div > fieldset > legend > h1"),
-    "Yes, my client wants to report their property income" -> By.id("value_0"),
-    "Remove" -> By.linkText("Remove"),
-    "Change country" -> By.cssSelector("#main-content > div > div > form > dl > div > dd.govuk-summary-list__actions.w-25 > ul > li:nth-child(1) > a"),
-    "No, don't add" -> By.id("addAnother-no"),
-    "Yes, add" -> By.id("addAnother"),
-    "Yes remove country" -> By.cssSelector("button.govuk-button.govuk-button--warning"),
-    "Country Option" -> By.id("country__option--0"),
-    "rent a room relief" -> By.id("ukAndForeignPropertyClaimExpensesOrRelief"),
-    "expenses" -> By.id("ukAndForeignPropertyClaimExpensesOrRelief-no"),
-    "use the property income allowance" -> By.id("ukAndForeignPropertyClaimPropertyIncomeAllowanceOrExpenses"),
-    "Expenses not PIA" -> By.id("ukAndForeignPropertyClaimPropertyIncomeAllowanceOrExpenses-no"),
-    "My Client does not want to report their property income" -> By.id("value_1"),
-
-    //UK & Foreign Property - UK property rentals and rent a room
-    "UK property rentals and rent a room About" -> By.id("rentals_and_rent_a_room_about_link"),
-    "About your property rental and rent a room" -> By.cssSelector("<h1 class=\"govuk-heading-l\">About your property rental and rent a room</h1>"),
-
-    "Rentals RAR HYFTS Yes" -> By.id("isRentalsRentARoomSectionComplete"),
-
-    //UK & Foreign Property - Income
-    "UK property rentals and rent a room Income" -> By.id("rentals_and_rent_a_room_income_link"),
-    "Income" -> By.cssSelector("<h1 class=\"govuk-heading-l\">Income</h1>"),
-
-    //UK & Foreign Property - Expenses
-    "UK property rentals and rent a room Expenses" -> By.id("rentals_and_rent_a_room_expenses_link"),
-
-    //UK & Foreign Property - Allowances
-    "UK property rentals and rent a room Allowances" -> By.id("rentals_and_rent_a_room_allowances_link"),
-
-    //UK & Foreign Property - Adjustments
-    "UK property rentals and rent a room Adjustments" -> By.id("rentals_and_rent_a_room_adjustments_link"),
-
-    //Foreign income - Dividends from foreign companies
-    "Dividends from foreign companies" -> By.id("foreign_income_dividends"),
-    "Foreign dividends by country" -> By.cssSelector("<h1 class=\"govuk-heading-l\">Foreign dividends by country</h1>"),
-    "Which country did you receive your dividend income from?" -> By.cssSelector("<h1 class=\"govuk-heading-l\">Which country did you receive your dividend income from?</h1>"),
-    "Which country did your client receive their dividend income from?Which country did your client receive their dividend income from?" -> By.cssSelector("<h1 class=\"govuk-heading-l\">Which country did your client receive their dividend income from?</h1>"),
-    "Tell us about dividend income you received from Portugal before foreign tax was deducted" -> By.cssSelector("<h1 class=\"govuk-heading-l\">Tell us about dividend income you received from Portugal before foreign tax was deducted</h1>"),
-    "Tell us about dividend income your client received from Italy before foreign tax was deducted" -> By.cssSelector("<h1 class=\"govuk-heading-l\">Tell us about dividend income your client received from Italy before foreign tax was deducted</h1>"),
-    "income received before foreign tax deducted amount" -> By.id("incomeBeforeForeignTaxDeducted"),
-    "Have you had foreign tax deducted from your dividend income from Portugal?" -> By.cssSelector("<h1 class=\"govuk-heading-l\">Have you had foreign tax deducted from your dividend income from Portugal?</h1>"),
-    "Has your client had foreign tax deducted from their dividend income from Italy?" -> By.cssSelector("<h1 class=\"govuk-heading-l\">Has your client had foreign tax deducted from their dividend income from Italy?</h1>"),
-    "How much foreign tax was deducted from your dividend income from Portugal?" -> By.cssSelector("<h1 class=\"govuk-heading-l\">How much foreign tax was deducted from your dividend income from Portugal?</h1>"),
-    "tax deducted from dividend income amount" -> By.id("howMuchForeignTaxDeductedFromDividendIncome"),
-    "How much foreign tax was deducted from your dividend income from Portugal?" -> By.cssSelector("<h1 class=\"govuk-heading-l\">How much foreign tax was deducted from your dividend income from Portugal?</h1>"),
-    "Your foreign dividends by country" -> By.cssSelector("<h1 class=\"govuk-heading-l\">Your foreign dividends by country</h1>"),
-    "Your client's foreign dividends by country" -> By.cssSelector("<h1 class=\"govuk-heading-l\">Your client's foreign dividends by country</h1>"),
-    "Have you completed this section?" -> By.cssSelector("<h1 class=\"govuk-heading-l\">Have you completed this section?</h1>"),
-
-    // Property - rent-a-room
-    "Rent-a-room" -> By.id("value_1"),
-    "Rentaroom About" -> By.id("rent_a_room_about_link"),
-    "Rent a room" -> By.cssSelector("#main-content > div > div > h1"),
-    "Jointly let rent a room" -> By.cssSelector("By.cssSelector(\"<h1 class=\"govuk-heading-l\">Jointly let rent a room</h1>\"),"),
-    "Jointly let rent a room, Yes" -> By.id("isJointlyLet"),
-    "Jointly let rent a room, No" -> By.id("isJointlyLet-no"),
-    "How much total rent a room income did you receive?" -> By.cssSelector("#main-content > div > div > form > div > h1 > label"),
-    "TotalRentaRoomAmt" -> By.id("totalIncomeAmount"),
-    "Claim expenses or rent a room relief" -> By.cssSelector("#main-content > div > div > form > h1"),
-    "Rent a room relief" -> By.id("rentARoomAmount"),
-    "Expenses" -> By.id("rentARoomAmount-no"),
-    "RentaRoomCYA About" -> By.cssSelector("#main-content > div > div > h1"),
-    "RaRAboutHYFTS,Yes" -> By.id("isRentARoomIsSectionComplete"),
-    "RaRAboutHYFTS,No" -> By.id("isRentARoomIsSectionComplete-no"),
-    "RaREpensesLink" -> By.id("rent_a_room_expenses_link"),
-    "Rent a room expenses" -> By.cssSelector("#main-content > div > div > h1"),
-    "Consolidated or individual expenses" -> By.cssSelector("#main-content > div > div > h1"),
-    "Claim consolidated expenses" -> By.id("isConsolidatedExpenses"),
-    "Claim individual expenses" -> By.id("isConsolidatedExpenses-no"),
-    "Claim consolidated expensesAmt"-> By.id("consolidatedExpensesAmount"),
-    "RaRExpensesHYFTS,Yes" -> By.id("isExpensesRRSectionComplete"),
-    "RaRExpensesHYFTS,No" -> By.id("isExpensesRRSectionComplete-no"),
-    "How much total income did your client get from all of their properties?" -> By.cssSelector("#main-content > div > div > h1"),
-    "Which rental property does your client have in the UK?" -> By.cssSelector("#main-content > div > div > h1"),
-    "How much total rent a room income did your client receive?" -> By.cssSelector("#main-content > div > div > h1"),
-
-    // Property, Pensions And Investments
-    "Income from property, pensions and investments tag"                   -> By.cssSelector("#task-list-3-status"),
-    "Income from property, pensions and investments"                       -> By.cssSelector("#main-content > div > div > ul > li:nth-child(3) > div.govuk-task-list__name-and-hint > a"),
-    "UK Property"                                                          -> By.id("RentalIncome"),
-    "Non UK Property"                                                      -> By.id("RentalIncome-2"),
-    "No income from property"                                              -> By.cssSelector("[value='NoRentalIncome']"),
-    "No income from pensions"                                              -> By.cssSelector("[value='NoIncomePensions']"),
-    "State Pension"                                                        -> By.id("Pensions"),
-    "Other UK pensions (occupational and private)"                         -> By.id("Pensions-2"),
-    "Unauthorised payments from pension schemes"                           -> By.id("Pensions-3"),
-    "Short service refunds from pensions"                                  -> By.id("Pensions-4"),
-    "Non UK pensions"                                                      -> By.id("Pensions-5"),
-    "No UK insurance gains"                                                -> By.cssSelector("[value='NoGains']"),
-    "Interest from UK banks and building societies"                        -> By.cssSelector("[value='fromUkBanks']"),
-    "Interest from UK trust funds or bonds"                                -> By.cssSelector("input#UkInterest-2"),
-    "Interest from gilt-edged or accrued income securities1"               -> By.id("UkInterest-3"),
-    "No income from UK interest"                                           -> By.cssSelector("[value='NoInterestIncome']"),
-    "Cash dividends from UK stocks and shares"                             -> By.cssSelector("input#UkDividendsSharesLoans"),
-    "Free or redeemable shares"                                            -> By.id("UkDividendsSharesLoans-4"),
-    "Stock dividends from UK companies"                                    -> By.id("UkDividendsSharesLoans-2"),
-    "Dividends from unit trusts or investment companies"                   -> By.id("UkDividendsSharesLoans-3"),
-    "Close company loans written off or released"                          -> By.id("UkDividendsSharesLoans-5"),
-    "No UK dividends, shares or loans"                                     -> By.cssSelector("[value='NoUkDividendsSharesLoans']"),
-    "Non-UK interest"                                                      -> By.cssSelector("[value='NonukInterest']"),
-    "Non-UK dividends while living in the UK"                              -> By.cssSelector("[value='NonukDividendsWhileLivingInTheUk]"),
-    "No non-UK interest, dividends or income from insurance policies"      -> By.cssSelector("[value='NoNonUkInterestDividendsInsurance']"),
-    "Life Insurance Option"                                                -> By.id("UkInsuranceGains"),
-    "Life Annuity"                                                         -> By.id("UkInsuranceGains-2"),
-    "Capital Redemption"                                                   -> By.id("UkInsuranceGains-3"),
-    "Voided ISA (a policy cancelled by your ISA manager)"                  -> By.id("UkInsuranceGains-4"),
-    "No UK insurance gains"                                                -> By.cssSelector("[value='NoGains']"),
-
-    //Payments into Pensions
-    "Payments into UK pensions"                                            -> By.id("PaymentsIntoPensions"),
-    "Payments into non-UK pensions"                                        -> By.id("PaymentsIntoPensions-2"),
-    "Annual allowances"                                                    -> By.id("PaymentsIntoPensions-3"),
-    "Overseas transfer charges1"                                            -> By.id("PaymentsIntoPensions-4"),
-    "Payments into UK pensions task list"                                  -> By.cssSelector("#main-content > div > div > ul > li:nth-child(4) > div.govuk-task-list__name-and-hint > a"),
-    "Payments into pensions tag"                                           -> By.cssSelector("#task-list-4-status"),
-
-    // Capital gains
-    "Trusts"                            -> By.cssSelector("[value='Trusts']"),
-    "Settlements"                       -> By.cssSelector("[value='Settlements']"),
-    "Estates of people who have died"   -> By.cssSelector("[value='Estates']"),
-    "No income from trusts"             -> By.cssSelector("value='NoIncome'"),
-
-    //Self Employment
-    "Self Employment" ->By.id("selfEmployment_link"),
-    "Review self-employments" -> By.cssSelector("#main-content > div > div > div.task-list > dl:nth-child(3) > div:nth-child(1) > dt > span > a"),
-    "Self Employment Save and continue" -> By.xpath("//*[contains(text(),'Save and continue')]"),
-    "Yes, I’ve completed this section" -> By.id("value"),
-    "No, I’ll come back to it later" -> By.id("value_1"),
-    "Industry sectors and working abroad" -> By.cssSelector("#main-content > div > div > div.task-list > dl:nth-child(3) > div:nth-child(2) > dt > span > a"),
-    "Self-employment Income" -> By.cssSelector("#main-content > div > div > div.task-list > dl:nth-child(3) > div:nth-child(3) > dt > span > a"),
-    "Amount-field" -> By.id("value"),
-    "Use the trading allowance" -> By.id("value_0"),
-    "Declare expenses" -> By.id("value_1"),
-    "first radio button" -> By.id("value_0"),
-    "second radio button" -> By.id("value_1"),
-    "third radio button" -> By.id("value_2"),
-    "Do you want to use trading allowance or declare expense" -> By.xpath("//a[substring(@href, string-length(@href) - string-length('income/change-trading-allowance') + 1) = 'income/change-trading-allowance']"),
-    "Self-employment Expenses categories" -> By.cssSelector("#main-content > div > div > div.task-list > dl:nth-child(3) > div:nth-child(4) > dt > span > a"),
-    "first option" -> By.id("value_0"),
-    "second option" -> By.id("value_1"),
-    "third option" -> By.id("value_2"),
-    "fifth option" -> By.id("value_4"),
-    "all checkbox options" -> By.cssSelector("#value_0, #value_1, #value_2, #value_3, #value_4, #value_5, #value_6, #value_7"),
-    "Office supplies" -> By.xpath("//dt[@class='govuk-summary-list__key']//a[substring(@href, string-length(@href) - string-length('expenses/office-supplies/amount') + 1) = 'expenses/office-supplies/amount']"),
-    "Goods to sell or use" -> By.xpath("//dt[@class='govuk-summary-list__key']//a[substring(@href, string-length(@href) - string-length('expenses/goods-sell-use/taxi-minicab-road-haulage-industry-Driver.instance') + 1) = 'expenses/goods-sell-use/taxi-minicab-road-haulage-industry-Driver.instance']"),
-    "Repairs and maintenance" -> By.xpath("//dt[@class='govuk-summary-list__key']//a[substring(@href, string-length(@href) - string-length('expenses/repairs-maintenance/amount') + 1) = 'expenses/repairs-maintenance/amount']"),
-    "Travel and accommodation for work" -> By.xpath("//dt[@class='govuk-summary-list__key']//a[substring(@href, string-length(@href) - string-length('expenses/travel/travel-accommodation-total-expenses') + 1) = 'expenses/travel/travel-accommodation-total-expenses']"),
-    "Advertising or marketing" -> By.xpath("//dt[@class='govuk-summary-list__key']//a[substring(@href, string-length(@href) - string-length('expenses/advertising-marketing/amount') + 1) = 'expenses/advertising-marketing/amount']"),
-    "Entertainment costs" -> By.xpath("//dt[@class='govuk-summary-list__key']//a[substring(@href, string-length(@href) - string-length('expenses/entertainment/disallowable-amount') + 1) = 'expenses/entertainment/disallowable-amount']"),
-    "Staff costs" -> By.xpath("//dt[@class='govuk-summary-list__key']//a[substring(@href, string-length(@href) - string-length('expenses/staff/amount') + 1) = 'expenses/staff/amount']"),
-    "Construction industry subcontractors" -> By.xpath("//dt[@class='govuk-summary-list__key']//a[substring(@href, string-length(@href) - string-length('expenses/construction-industry/amount') + 1) = 'expenses/construction-industry/amount']"),
-    "Professional fess" -> By.xpath("//dt[@class='govuk-summary-list__key']//a[substring(@href, string-length(@href) - string-length('expenses/professional-fees/amount') + 1) = 'expenses/professional-fees/amount']"),
-    "Interest paid on bank and other business loans" -> By.xpath("//dt[@class='govuk-summary-list__key']//a[substring(@href, string-length(@href) - string-length('expenses/interest-bank-interest-loans/amount') + 1) = 'expenses/interest-bank-business-loans/amount']"),
-    "Bank, credit card and other financial charges" -> By.xpath("//dt[@class='govuk-summary-list__key']//a[substring(@href, string-length(@href) - string-length('expenses/bank-credit-card-financial-charges/amount') + 1) = 'expenses/bank-credit-card-financial-charges/amount']"),
-    "Irrecoverable debt" -> By.xpath("//dt[@class='govuk-summary-list__key']//a[substring(@href, string-length(@href) - string-length('expenses/irrecoverable-debts/amount') + 1) = 'expenses/irrecoverable-debts/amount']"),
-    "Depreciation costs" -> By.xpath("//dt[@class='govuk-summary-list__key']//a[substring(@href, string-length(@href) - string-length('expenses/depreciation/disallowable-amount') + 1) = 'expenses/depreciation/disallowable-amount']"),
-    "Other expenses" -> By.xpath("//dt[@class='govuk-summary-list__key']//a[substring(@href, string-length(@href) - string-length('expenses/other-expenses/amount') + 1) = 'expenses/other-expenses/amount']"),
-    "Workplace running costs (including working from home)" -> By.xpath("//dt[@class='govuk-summary-list__key']//a[substring(@href, string-length(@href) - string-length('expenses/workplace-running-costs/working-from-home/more-than-25-hours') + 1) = 'expenses/workplace-running-costs/working-from-home/more-than-25-hours']"),
-    "25 - 50 hours" -> By.id("value25To50"),
-    "51 - 100 hours" -> By.id("value51To100"),
-    "101 hours or more" -> By.id("value101Plus"),
-    "Actual costs" -> By.id("value_1"),
-    "1 person" -> By.id("onePerson"),
-    "2 people" -> By.id("twoPeople"),
-    "3 or more people" -> By.id("threePeople"),
-    "Allowances for vehicles, equipment and buildings" -> By.xpath("//a[text()=' Allowances for vehicles, equipment and buildings ']"),
-    "Zero-emission car allowance" -> By.xpath("//input[@value='zeroEmissionCar']"),
-    "Zero-emission goods vehicle allowance" -> By.xpath("//input[@value='zeroEmissionGoodsVehicle']"),
-    "Electric vehicle charge-point allowance" -> By.xpath("//input[@value='electricVehicleChargepoint']"),
-    "Structures and buildings allowance" -> By.xpath("//input[@value='structuresAndBuildings']"),
-    "Freeport structures and buildings allowance" -> By.xpath("//input[@value='specialTaxSitesStructuresAndBuildings']"),
-    "Annual investment allowance (business equipment)" -> By.xpath("//input[@value='annualInvestment']"),
-    "'Writing down' allowance" -> By.xpath("//a[text()=' ‘Writing down’ allowance ']"),
-    "textbox" -> By.id("value"),
-    "Car or goods vehicle" -> By.id("value_0"),
-    "Flat rate" -> By.id("value_0"),
-    "Balancing allowance (allowance on sales)" -> By.xpath("//input[@value='balancing']"),
-    "Balancing charge" -> By.xpath("//a[text()=' Balancing charge ']"),
-    "Balancing allowance" -> By.xpath("//a[text()=' Balancing allowance ']"),
-    "Zero-emission cars" -> By.xpath("//a[text()=' Zero-emission cars ']"),
-    "10 percent" -> By.id("10%"),
-    "full cost" -> By.id("fullCost"),
-    "Zero-emission goods vehicles" -> By.xpath("//a[text()=' Zero-emission goods vehicles ']"),
-    "Electric vehicle charge-points" -> By.xpath("//a[text()=' Electric vehicle charge-points ']"),
-    "Structures and buildings allowance" -> By.xpath("//a[text()=' Structures and buildings allowance ']"),
-    "QualifyingUse Day" -> By.id("structuresBuildingsQualifyingUseDate.day"),
-    "QualifyingUse Month" -> By.id("structuresBuildingsQualifyingUseDate.month"),
-    "QualifyingUse Year" -> By.id("structuresBuildingsQualifyingUseDate.year"),
-    "Self employment building number" -> By.id("buildingNumber"),
-    "Self employment postcode" -> By.id("postcode"),
-    "Freeport sites" -> By.xpath("//a[text()=' Freeport sites ']"),
-    "Construction start Day" -> By.id("contractStartDate.day"),
-    "Construction start Month" -> By.id("contractStartDate.month"),
-    "Construction start Year" -> By.id("contractStartDate.year"),
-    "Qualifying start Day" -> By.id("qualifyingUseStartDate.day"),
-    "Qualifying start Month" -> By.id("qualifyingUseStartDate.month"),
-    "Qualifying start Year" -> By.id("qualifyingUseStartDate.year"),
-    "Annual investment allowance" -> By.xpath("//a[text()=' Annual investment allowance ']"),
-    "Writing down allowance" -> By.cssSelector("#main-content > div > div > h1"),
-    "Work out your adjusted profit or loss" -> By.xpath("//a[text()=' Work out your adjusted profit or loss ']"),
-    "When you reported the loss" -> By.cssSelector("#main-content > div > div > form > h1")
-  )
-
-  val elements: Map[String, By] = Map(commonMapping: _*)
-
-
-}
